Collection Completed changes
Bot will now PM user when they complete a collection. It sets collection as completed without an amount so it can be listed in ->col -completed, and displays have been changed to not display when clout amount is 0/undefined.

Forge shows when you get a duplicate card again.

Re-added dif as an alias to diff

Minor spelling/grammar changes in hero