Fixes col reset, adds ->trans auction
Made a fix for col reset adding a null for number if it is your first reset. Re-worded help for query operators, achievement now looks for a col reset of amount >0, instead of length >0. Profile now lists completed cols and clouted cols at different times, in line with the new completed definition