Remove -diff filtering from withGlobalCards
This caused command reply blocking for periods of time relative to how many cards a user had.

Very few commands have a need for this so the diff filtering was moved to within those commands.

Also a minor fix in auc info all, allowing you to query diff and bid